requirements.txt
Flask
APScheduler
pytz
instagrapi
instaloader
praw
pillow
requests
python-dotenv
moviepy